backend/requirements.txt
fastapi
uvicorn
httpx
pydantic
psycopg2-binary
redis
openai>=1.23.0
pytest
